Feature/sunset sunrise in config get

+++ WordClock/Clock.h
@@ -7,7 +7,7 @@
 #include <Wire.h> // Connected via I2C-bus on pins D1 (GPIO5/SCL) and D2 (GPIO4/SDA)
 
 struct Time {
-  uint8_t year;
+  uint16_t year;
   uint8_t month;
   uint8_t day;
   uint8_t hour;
@@ -23,7 +23,7 @@ class Clock {
     void setTime(Time time);
     Time getTime();
   private:
-    uint8_t getYear();
+    uint16_t getYear();
     uint8_t getMonth();
     uint8_t getDate();
     uint8_t getHour();
